+++ test/AutoRelinkSpec.hs
@@ -39,7 +39,7 @@ import Database.Loader (LoadOptions (..))
 import Database.Manager (CachePolicy (..), LoadSource (..), RawLoad (..), loadDatabaseRawWithCrossDB)
 import SynonymDB (emptySynonymDB)
 import Types (AllocationKey (..), GeographyPolicy (..))
-import UnitConversion (UnitConfig, UnitDef (..), defaultUnitConfig, mkUnitConfig, ucDimensionOrder, ucOriginalKeys, ucUnits)
+import UnitConversion (UnitConfig, UnitDef (..), defaultUnitConfig, mkUnitConfig, ucDimensionOrder, ucUnits)
 
 {- | Copy regular files from one directory into another (non-recursive,
 which is all the EcoSpold v2 fixtures here need).
@@ -85,7 +85,6 @@ withGram =
     mkUnitConfig
         (ucDimensionOrder defaultUnitConfig)
         (M.insert "g" (UnitDef [1, 0, 0, 0, 0, 0, 0, 0] 0.001) (ucUnits defaultUnitConfig))
-        (M.insert "g" "g" (ucOriginalKeys defaultUnitConfig))
 
 spec :: Spec
 spec = do